+++ source/solvers/fluid_dynamics_matrix_based.cc
@@ -1517,8 +1517,8 @@ FluidDynamicsMatrixBased<dim>::solve_system_GMRES(
 
   const AffineConstraints<double> &constraints_used =
     initial_step ? nonzero_constraints : zero_constraints_used;
-  const double normalize_metric = this->get_residual_normalization_metric();
-  const double current_residual = this->system_rhs.l2_norm() / normalize_metric;
+  const double rescale_metric   = this->get_residual_rescale_metric();
+  const double current_residual = this->system_rhs.l2_norm() / rescale_metric;
   const double linear_solver_tolerance =
     std::max(relative_residual * current_residual, absolute_residual);
   if (this->simulation_parameters.linear_solver.at(PhysicsID::fluid_dynamics)
@@ -1527,15 +1527,15 @@ FluidDynamicsMatrixBased<dim>::solve_system_GMRES(
       this->pcout << "  -Tolerance of iterative solver is : "
                   << linear_solver_tolerance << std::endl;
     }
-  const double non_normalized_linear_solver_tolerance =
-    linear_solver_tolerance * normalize_metric;
+  const double non_rescaled_linear_solver_tolerance =
+    linear_solver_tolerance * rescale_metric;
   GlobalVectorType completely_distributed_solution(this->locally_owned_dofs,
                                                    this->mpi_communicator);
 
   SolverControl solver_control(this->simulation_parameters.linear_solver
                                  .at(PhysicsID::fluid_dynamics)
                                  .max_iterations,
-                               non_normalized_linear_solver_tolerance,
+                               non_rescaled_linear_solver_tolerance,
                                true,
                                true);
   bool          extra_verbose = false;
@@ -1603,8 +1603,7 @@ FluidDynamicsMatrixBased<dim>::solve_system_GMRES(
                 this->pcout
                   << "  -Iterative solver took : " << solver_control.last_step()
                   << " steps to reach a residual norm of "
-                  << solver_control.last_value() / normalize_metric
-                  << std::endl;
+                  << solver_control.last_value() / rescale_metric << std::endl;
               }
           }
 
@@ -1667,8 +1666,8 @@ FluidDynamicsMatrixBased<dim>::solve_system_BiCGStab(
   const AffineConstraints<double> &constraints_used =
     initial_step ? nonzero_constraints : zero_constraints_used;
 
-  const double normalize_metric = this->get_residual_normalization_metric();
-  const double current_residual = this->system_rhs.l2_norm() / normalize_metric;
+  const double rescale_metric   = this->get_residual_rescale_metric();
+  const double current_residual = this->system_rhs.l2_norm() / rescale_metric;
   const double linear_solver_tolerance =
     std::max(relative_residual * current_residual, absolute_residual);
   if (this->simulation_parameters.linear_solver.at(PhysicsID::fluid_dynamics)
@@ -1677,8 +1676,8 @@ FluidDynamicsMatrixBased<dim>::solve_system_BiCGStab(
       this->pcout << "  -Tolerance of iterative solver is : "
                   << linear_solver_tolerance << std::endl;
     }
-  const double non_normalized_linear_solver_tolerance =
-    linear_solver_tolerance * normalize_metric;
+  const double non_rescaled_linear_solver_tolerance =
+    linear_solver_tolerance * rescale_metric;
   GlobalVectorType completely_distributed_solution(this->locally_owned_dofs,
                                                    this->mpi_communicator);
 
@@ -1692,7 +1691,7 @@ FluidDynamicsMatrixBased<dim>::solve_system_BiCGStab(
   SolverControl solver_control(this->simulation_parameters.linear_solver
                                  .at(PhysicsID::fluid_dynamics)
                                  .max_iterations,
-                               non_normalized_linear_solver_tolerance,
+                               non_rescaled_linear_solver_tolerance,
                                true,
                                true);
   TrilinosWrappers::SolverBicgstab solver(solver_control, solver_parameters);
@@ -1730,8 +1729,7 @@ FluidDynamicsMatrixBased<dim>::solve_system_BiCGStab(
                 this->pcout
                   << "  -Iterative solver took : " << solver_control.last_step()
                   << " steps to reach a residual norm of "
-                  << solver_control.last_value() / normalize_metric
-                  << std::endl;
+                  << solver_control.last_value() / rescale_metric << std::endl;
               }
             constraints_used.distribute(completely_distributed_solution);
             this->newton_update = completely_distributed_solution;
@@ -1773,12 +1771,12 @@ FluidDynamicsMatrixBased<dim>::solve_system_direct(
 
   const AffineConstraints<double> &constraints_used =
     initial_step ? nonzero_constraints : zero_constraints_used;
-  const double normalize_metric = this->get_residual_normalization_metric();
-  const double current_residual = this->system_rhs.l2_norm() / normalize_metric;
+  const double rescale_metric   = this->get_residual_rescale_metric();
+  const double current_residual = this->system_rhs.l2_norm() / rescale_metric;
   const double linear_solver_tolerance =
     std::max(relative_residual * current_residual, absolute_residual);
-  const double non_normalized_linear_solver_tolerance =
-    linear_solver_tolerance * normalize_metric;
+  const double non_rescaled_linear_solver_tolerance =
+    linear_solver_tolerance * rescale_metric;
 
   GlobalVectorType completely_distributed_solution(this->locally_owned_dofs,
                                                    this->mpi_communicator);
@@ -1786,7 +1784,7 @@ FluidDynamicsMatrixBased<dim>::solve_system_direct(
   SolverControl solver_control(this->simulation_parameters.linear_solver
                                  .at(PhysicsID::fluid_dynamics)
                                  .max_iterations,
-                               non_normalized_linear_solver_tolerance,
+                               non_rescaled_linear_solver_tolerance,
                                true,
                                true);
   TrilinosWrappers::SolverDirect solver(solver_control);