+++ source/solvers/vof_algebraic_interface_reinitialization.cc
@@ -621,12 +621,12 @@ VOFAlgebraicInterfaceReinitialization<dim>::solve_linear_system(
       .relative_residual;
 
   // Set linear solver tolerance
-  const double normalize_metric = this->get_residual_normalization_metric();
-  const double current_residual = this->system_rhs.l2_norm() / normalize_metric;
+  const double rescale_metric   = this->get_residual_rescale_metric();
+  const double current_residual = this->system_rhs.l2_norm() / rescale_metric;
   const double linear_solver_tolerance =
     std::max(relative_residual * current_residual, absolute_residual);
-  const double non_normalized_linear_solver_tolerance =
-    linear_solver_tolerance * normalize_metric;
+  const double non_rescaled_linear_solver_tolerance =
+    linear_solver_tolerance * rescale_metric;
 
   if (verbose)
     {
@@ -661,7 +661,7 @@ VOFAlgebraicInterfaceReinitialization<dim>::solve_linear_system(
     this->simulation_parameters.vof_subequations_linear_solvers
       .at(VOFSubequationsID::algebraic_interface_reinitialization)
       .max_iterations,
-    non_normalized_linear_solver_tolerance,
+    non_rescaled_linear_solver_tolerance,
     true,
     true);
 
@@ -682,8 +682,7 @@ VOFAlgebraicInterfaceReinitialization<dim>::solve_linear_system(
     {
       this->pcout << "  -Iterative solver took: " << solver_control.last_step()
                   << " steps to reach a residual norm of "
-                  << solver_control.last_value() / normalize_metric
-                  << std::endl;
+                  << solver_control.last_value() / rescale_metric << std::endl;
     }
 
   // Update constraints and newton vectors