+++ tests/core/non_linear_test_system_01.h
@@ -139,7 +139,7 @@ public:
     return dummy_constraints;
   };
   virtual double
-  get_residual_normalization_metric() const override
+  get_residual_rescale_metric() const override
   {
     return 1.0;
   }